requirements.txt
streamlit==0.73.1
tensorflow-cpu==2.4.0
numpy==1.19.4
Keras==2.4.3
Pillow==8.0.1
